requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
pytesseract==0.3.10
openai==0.28.0
Pillow==9.5.0
